test: move test files back to original position
